pyproject.toml
[project]
name = "ice-breaker"
version = "0.1.0"
description = "This is a web application crawling Linkedin & Twitter data about a person and customizes an ice breaker with them."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.14",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.0",
    "tweepy>=4.15.0",
]
